+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Remove support for platform version `221.*`
 
 ### Fixed
+- Fix error in 3D preview of a line without texture coordinates
 - Wrap file operations in read-actions
 - Replace usages of internal method `StartupManager.runAfterOpened()` with a `StartupActivity`
 - Override `getActionUpdateThread()` in existing implementations of `AnAction`